refactor: Divide wdcuration.wdcuration modules
- Create 5 new modules
	- api_searches for functions related to API queries
	- quickstatements for functions related to quickstatements
	- sparql for sparql query related functions
	- utils for general utilities
	- dicthandler with functions for updating/creating
	dictionaries
- Divide test suite into separate test suites, as per the division seen
  above.
- Refactors imports in __init__.py

- Closes #16

@lubianat Let me know if you agree with this way of dividing wdcuration into
separate submodules, I tried to divide them in a way that made sense to
me but you might reorganize them another way.
